+++ package/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Next
+
+### What's Changed
+
+- better DX for importing constants & presets
+
 ## 0.2.0 (2026-05-29)
 
 ### What's Changed